# frontend/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ArtifactFlow API",
    "description": "Multi-agent system API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/skills": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Admin List Skills",
        "description": "\u5217\u51fa\u6240\u6709 shared skill,\u4e0d\u6309 admin \u81ea\u5df1\u7684\u90e8\u95e8\u53ef\u89c1\u6027\u8fc7\u6ee4\u3002",
        "operationId": "admin_list_skills_api_v1_admin_skills_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSkillListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/skills/{skill_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Admin Get Skill Detail",
        "description": "\u8bfb\u53d6\u5171\u4eab skill \u6b63\u6587\uff0c\u4e0d\u6309\u5f53\u524d\u7ba1\u7406\u5458\u81ea\u5df1\u7684\u90e8\u95e8\u53ef\u89c1\u6027\u8fc7\u6ee4\u3002",
        "operationId": "admin_get_skill_detail_api_v1_admin_skills__skill_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Skill Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Admin Update Skill",
        "description": "\u7f16\u8f91 dynamic shared skill \u7684 visibility/default_enabled\u3002seeded \u4ecd config-owned\u3002",
        "operationId": "admin_update_skill_api_v1_admin_skills__skill_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Skill Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminSkillUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSkillItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Admin Delete Skill",
        "description": "\u5220\u9664\u4efb\u610f dynamic skill(\u7ed5\u8fc7\u53ef\u89c1\u6027;seeded \u2192 400)\u3002\u7ea7\u8054\u6e05 user_skill/dept \u89c4\u5219\u3002",
        "operationId": "admin_delete_skill_api_v1_admin_skills__skill_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Skill Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/skills/import": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Admin Import Skill",
        "description": "\u5bfc\u5165\u5171\u4eab skill(owner=null),\u53ef\u6307\u5b9a visibility/default_enabled \u521d\u59cb\u503c\u3002",
        "operationId": "admin_import_skill_api_v1_admin_skills_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_admin_import_skill_api_v1_admin_skills_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/skills/{skill_id}/export": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Admin Export Skill",
        "description": "\u5bfc\u51fa shared catalog skill,\u4e0d\u6309\u5f53\u524d admin \u7684\u90e8\u95e8\u53ef\u89c1\u6027\u8fc7\u6ee4\u3002",
        "operationId": "admin_export_skill_api_v1_admin_skills__skill_id__export_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Skill Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/department-access/{dept_id}/skills/{slug}": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Put Department Skill Rule",
        "operationId": "put_department_skill_rule_api_v1_admin_department_access__dept_id__skills__slug__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dept_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dept Id"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete Department Skill Rule",
        "operationId": "delete_department_skill_rule_api_v1_admin_department_access__dept_id__skills__slug__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dept_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dept Id"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/skills": {
      "get": {
        "tags": [
          "skills"
        ],
        "summary": "List Skills",
        "description": "\u5217\u51fa\u5bf9\u5f53\u524d\u7528\u6237\u53ef\u89c1\u7684 skill + \u6709\u6548\u542f\u7528\u6001\u3002",
        "operationId": "list_skills_api_v1_skills_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/skills/{skill_id}": {
      "get": {
        "tags": [
          "skills"
        ],
        "summary": "Get Skill Detail",
        "description": "\u6309\u9700\u8bfb\u53d6\u5f53\u524d\u7528\u6237\u53ef\u89c1 skill \u7684 SKILL.md \u6b63\u6587\uff1b\u4e0d\u53ef\u89c1\u4ecd\u8fd4\u56de 404\u3002",
        "operationId": "get_skill_detail_api_v1_skills__skill_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Skill Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "skills"
        ],
        "summary": "Delete Skill",
        "description": "\u5220\u9664\u81ea\u5df1\u5bfc\u5165\u7684 dynamic skill\u3002\u4e0d\u53ef\u89c1 \u2192 404;seeded \u2192 400;\u975e\u672c\u4eba\u5171\u4eab \u2192 403\u3002",
        "operationId": "delete_skill_api_v1_skills__skill_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Skill Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/skills/import": {
      "post": {
        "tags": [
          "skills"
        ],
        "summary": "Import Skill",
        "description": "\u5bfc\u5165\u79c1\u6709 skill zip(owner=\u672c\u4eba,\u7acb\u5373\u8fdb\u81ea\u5df1\u7684 L1)\u3002\u786c\u95e8\u62d2\u6536 \u2192 422 \u7ed3\u6784\u5316\nfindings;\u8d85\u5355\u5305\u4e0a\u9650 \u2192 422;\u8d85\u5b58\u50a8\u914d\u989d \u2192 413;\u4e2a\u4eba\u6570\u91cf\u8fbe\u4e0a\u9650\u6216 slug \u649e\u540d \u2192 409\u3002",
        "operationId": "import_skill_api_v1_skills_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_skill_api_v1_skills_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/skills/{skill_id}/export": {
      "get": {
        "tags": [
          "skills"
        ],
        "summary": "Export Skill",
        "description": "\u5bfc\u51fa DB \u4e2d\u4fdd\u5b58\u7684 skill zip\u3002\u4e0d\u53ef\u89c1 \u2192 404\u3002",
        "operationId": "export_skill_api_v1_skills__skill_id__export_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Skill Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/skills/{skill_id}/enabled": {
      "put": {
        "tags": [
          "skills"
        ],
        "summary": "Set Skill Enabled",
        "description": "\u4e2a\u4eba\u5f00\u5173\u67d0 skill \u662f\u5426\u8fdb L1 \u7d22\u5f15(\u5199 user_skill \u8986\u76d6)\u3002",
        "operationId": "set_skill_enabled_api_v1_skills__skill_id__enabled_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Skill Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillToggleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminSkillItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Stable skill id used by management APIs"
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "description": "Shared-catalog slug"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Display name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "One-line description"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "department"
            ],
            "title": "Visibility",
            "description": "public (default allow) or department (default deny)"
          },
          "default_enabled": {
            "type": "boolean",
            "title": "Default Enabled",
            "description": "Whether this shared skill enters users' L1 index by default"
          },
          "source": {
            "type": "string",
            "enum": [
              "seeded",
              "dynamic"
            ],
            "title": "Source",
            "description": "seeded skills are config-owned; dynamic shared skills are editable"
          },
          "has_extra_files": {
            "type": "boolean",
            "title": "Has Extra Files",
            "description": "Whether the skill bundle contains files beyond SKILL.md"
          },
          "can_edit": {
            "type": "boolean",
            "title": "Can Edit",
            "description": "True only for dynamic shared skills"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "visibility",
          "default_enabled",
          "source",
          "has_extra_files",
          "can_edit"
        ],
        "title": "AdminSkillItem",
        "description": "Admin shared skill catalog item (not filtered by the admin user's department)."
      },
      "AdminSkillListResponse": {
        "properties": {
          "skills": {
            "items": {
              "$ref": "#/components/schemas/AdminSkillItem"
            },
            "type": "array",
            "title": "Skills",
            "description": "All shared skills, including seeded read-only ones"
          }
        },
        "type": "object",
        "required": [
          "skills"
        ],
        "title": "AdminSkillListResponse",
        "description": "GET /api/v1/admin/skills response."
      },
      "AdminSkillUpdateRequest": {
        "properties": {
          "visibility": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "public",
                  "department"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Visibility",
            "description": "New shared visibility. Changing it clears department rules."
          },
          "default_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Enabled",
            "description": "New default L1 enabled state for users without an override."
          }
        },
        "type": "object",
        "title": "AdminSkillUpdateRequest",
        "description": "PATCH /api/v1/admin/skills/{skill_id} request body."
      },
...
      "Body_admin_import_skill_api_v1_admin_skills_import_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "department"
            ],
            "title": "Visibility",
            "default": "public"
          },
          "default_enabled": {
            "type": "boolean",
            "title": "Default Enabled",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "file",
          "visibility",
          "default_enabled"
        ],
        "title": "Body_admin_import_skill_api_v1_admin_skills_import_post"
      },
...
      "Body_import_skill_api_v1_skills_import_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_skill_api_v1_skills_import_post"
      },
...
      "FindingItem": {
        "properties": {
          "rule": {
            "type": "string",
            "title": "Rule",
            "description": "Stable rule id, e.g. 'zip.invalid' / 'md.body_empty'."
          },
          "severity": {
            "type": "string",
            "title": "Severity",
            "description": "'error' (rejected) or 'warning' (surfaced only)."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable explanation."
          }
        },
        "type": "object",
        "required": [
          "rule",
          "severity",
          "message"
        ],
        "title": "FindingItem",
        "description": "\u4e00\u6761 validator finding\uff1brule id \u7a33\u5b9a\uff0c\u524d\u7aef\u6309 severity \u6e32\u67d3\u3002"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError",
        "required": [
          "detail"
        ]
      },
...
      "SkillDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Stable skill id"
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "description": "User-context skill slug"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Display name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "One-line description"
          },
          "skill_md": {
            "type": "string",
            "title": "Skill Md",
            "description": "SKILL.md guidance body with YAML frontmatter removed"
          },
          "source": {
            "type": "string",
            "enum": [
              "seeded",
              "dynamic"
            ],
            "title": "Source",
            "description": "Config-owned or UI-imported source"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "private",
              "public",
              "department"
            ],
            "title": "Visibility",
            "description": "Effective catalog visibility"
          },
          "has_extra_files": {
            "type": "boolean",
            "title": "Has Extra Files",
            "description": "Whether the downloadable bundle contains files beyond SKILL.md"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "skill_md",
          "source",
          "visibility",
          "has_extra_files"
        ],
        "title": "SkillDetailResponse",
        "description": "On-demand SKILL.md preview for one management-list item."
      },
      "SkillImportResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Always 'imported' on success."
          },
          "skill": {
            "$ref": "#/components/schemas/SkillItem",
            "description": "The imported skill as it appears in the list."
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingItem"
            },
            "type": "array",
            "title": "Findings",
            "description": "Non-blocking warnings surfaced by the validator (may be empty)."
          }
        },
        "type": "object",
        "required": [
          "status",
          "skill",
          "findings"
        ],
        "title": "SkillImportResponse",
        "description": "POST /api/v1/skills/import(+ admin \u53d8\u4f53)\u6210\u529f\u54cd\u5e94\u3002\u786c\u95e8\u62d2\u6536\u8d70 422,\ndetail = {message, findings[]} \u540c\u4e00 FindingItem \u5f62\u72b6\u3002"
      },
      "SkillItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Stable skill id used by management APIs"
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "description": "User-context skill slug"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Display name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "One-line description (the L1 index text)"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "Effective enabled state (config default overridden by the user's setting). Controls whether the skill enters the model's L1 <available_skills> index. A disabled skill is still visible and can be activated on demand via its button."
          },
          "default_enabled": {
            "type": "boolean",
            "title": "Default Enabled",
            "description": "Config-seed default; shown so the UI can flag skills the user overrode."
          },
          "is_overridden": {
            "type": "boolean",
            "title": "Is Overridden",
            "description": "Whether the user has an explicit personal enable/disable for this skill."
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "'seeded' (config-owned, read-only in the UI) or 'dynamic' (imported via the UI; deletable by its owner or an admin)."
          },
          "has_extra_files": {
            "type": "boolean",
            "title": "Has Extra Files",
            "description": "Whether the skill bundle contains files beyond SKILL.md. Controls whether read_skill points to mount_skill."
          },
          "visibility": {
            "type": "string",
            "title": "Visibility",
            "description": "private (owner-only) | public (shared) | department"
          },
          "is_owner": {
            "type": "boolean",
            "title": "Is Owner",
            "description": "Whether the current user imported (owns) this dynamic skill."
          },
          "shadowed_by_private": {
            "type": "boolean",
            "title": "Shadowed By Private",
            "description": "Whether this shared skill is currently hidden at runtime by the user's same-slug private skill."
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "enabled",
          "default_enabled",
          "is_overridden",
          "source",
          "has_extra_files",
          "visibility",
          "is_owner",
          "shadowed_by_private"
        ],
        "title": "SkillItem",
        "description": "\u4e00\u4e2a\u5bf9\u7528\u6237\u53ef\u89c1\u7684 skill + \u5176\u6709\u6548\u542f\u7528\u6001\u3002"
      },
      "SkillListResponse": {
        "properties": {
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillItem"
            },
            "type": "array",
            "title": "Skills",
            "description": "Skills visible to the user, with effective enabled state."
          }
        },
        "type": "object",
        "required": [
          "skills"
        ],
        "title": "SkillListResponse",
        "description": "GET /api/v1/skills response."
      },
      "SkillToggleRequest": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "Personal enable/disable override for this skill."
          }
        },
        "type": "object",
        "required": [
          "enabled"
        ],
        "title": "SkillToggleRequest",
        "description": "PUT /api/v1/skills/{skill_id}/enabled request body."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type",
          "input",
          "ctx"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}